One-for-All-129 - Implemented the convertValueSelectBox.
