+++ apps/web/src/feature/record/import/convert-records/convert-value-selectbox.ts
@@ -1,0 +1,18 @@
+import type { FieldOptionsSelectBox } from "@feature/app/schema";
+
+export const convertValueSelectBox = (
+  originalValue: string,
+  options: FieldOptionsSelectBox,
+) => {
+  const foundSelector = options.selector.find(
+    ({ label }) => label === originalValue,
+  );
+
+  const errorMessage = foundSelector ? undefined : "選択肢が見つかりません";
+
+  return {
+    originalValue,
+    value: foundSelector?.value ?? originalValue,
+    errorMessage,
+  };
+};
