+++ apps/web/src/feature/record/import/convert-records/convert-value.ts
@@ -1,4 +1,5 @@
 import type { Field } from "@feature/app/schema";
+import { convertValueSelectBox } from "@feature/record/import/convert-records/convert-value-selectbox";
 
 export const convertValue = (
   originalValue: string,
@@ -11,31 +12,14 @@ export const convertValue = (
   switch (field.fieldKind) {
     case "text":
     case "multipleText":
+    case "lookup":
       return {
         originalValue: undefined,
         value: originalValue,
         errorMessage: undefined,
       };
     case "selectBox": {
-      const options = field.options;
-      const foundSelector = options.selector.find(
-        ({ label }) => label === originalValue,
-      );
-
-      const errorMessage = foundSelector ? undefined : "選択肢が見つかりません";
-
-      return {
-        originalValue,
-        value: foundSelector?.value ?? originalValue,
-        errorMessage,
-      };
-    }
-    case "lookup": {
-      return {
-        originalValue: undefined,
-        value: originalValue,
-        errorMessage: undefined,
-      };
+      return convertValueSelectBox(originalValue, field.options);
     }
     case "date": {
       return {
